+++ src/main/java/org/opensearch/security/OpenSearchSecurityPlugin.java
@@ -677,7 +677,8 @@ public final class OpenSearchSecurityPlugin extends OpenSearchSecuritySSLPlugin
                         sslSettingsManager,
                         Objects.requireNonNull(userService),
                         sslCertReloadEnabled,
-                        passwordHasher
+                        passwordHasher,
+                        rsIndexHandler
                     )
                 );